REF: Make new "multi" functions device-agnostic
